Fix script logic so it dumps the literal string to PATH
The export PATH command was being interpolated before being inserted in
~/.zshrc.

As a result, if any PATHs were being updated before this script, it
would be lost due to the hardcoded PATH caused by the interpolation.

This fixes it by forcing no interpolation when adding the command to the
path.